+++ src/galgeleg/Galgeklient.java
@@ -2,6 +2,8 @@ package galgeleg;
 
 import java.rmi.Naming;
 import java.util.ArrayList;
+import java.util.logging.Level;
+import java.util.logging.Logger;
 
 public class Galgeklient {
 
@@ -10,7 +12,12 @@ public class Galgeklient {
         GalgeI spil = (GalgeI) Naming.lookup("rmi://localhost:1099/galgeleg");
 
         spil.nulstil();
-
+     try {
+          spil.hentOrdFraDr();
+      } catch (Exception ex) {
+          Logger.getLogger(Galgelogik.class.getName()).log(Level.SEVERE, null, ex);
+      }
+    
         spil.logStatus();
 
         spil.gætBogstav("e");
